backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
Pillow==10.3.0
tensorflow==2.17.0
python-dotenv
langchain